Create MatrixOperations.java
With this class we can find determinant of square matrix. It need a control that matrix is really have equal number of rows and columns.
If matrix is 2x2 or 3x3 code works fast, else we use recursion (it isn't good variant, and I think about eliminating of recursion.
Code based on code of author Ocheredko.